==> 03-lecturer-variation.Rmd <==
---
title: "Lecturer Questions: Variation in number and category"
author: "George Kinnear"
date: "2024-04-04"
output: 
  html_document:
    code_folding: hide
    df_print: paged
---

```{r setup, message=FALSE, warning=FALSE}
library(tidyverse)
library(janitor)

# Tables
library(knitr)
library(kableExtra)
basic_kable = function(df, ...) {
  df %>% 
    kable(...) %>%
    kable_styling(bootstrap_options = "striped", full_width = F)
}

# Plot styling
theme_set(theme_minimal())
```

Load data:

```{r read-data, message=FALSE, warning=FALSE}
all_qs_all_lecs <- read_csv("data/ANON_LQ_Maths_All.csv")
anon_courses <- read_csv("data/ANON_courses.csv")

codes <- c("A", "C", "E", "F", "N", "O", "P", "U", "W", "X")
# Collapsing C, P, X, O into Other
collapsed_codes <- c("A", "O", "E", "F", "N", "O", "O", "U", "W", "O")

code_names <- c("Admin", "Convention", "Evaluation", "Fact", "Next step", "Other",
                "Proof framework", "Understanding check", "Warrant", "Example generation")

code_details <- tibble(code = codes, code_name = code_names) %>% 
  mutate(
    collapsed_code = case_when(
      code %in% c("C", "P", "X", "O") ~ "Other",
      TRUE ~ code_name
    )
  ) %>% 
  add_row(
    code = "Total",
    code_name = "Total",
    collapsed_code = "Total"
  )

all_qs_advanced <- all_qs_all_lecs %>% 
  # restrict to only those courses that were deemed "advanced mathematics"
  semi_join(
    anon_courses %>% filter(advanced == "yes"),
    by = c("course" = "anon_code")
  )

# Making list of all lectures and category code combos
lec_cat <- expand_grid(
  all_qs_all_lecs %>% select(course, session) %>% distinct(),
  category = codes
)
```


```{r message=FALSE, warning=FALSE}
add_group_totals <- function(df) {
  df %>% 
    # Add rows containing the totals
    bind_rows(
      df %>%
        group_by(course, session) %>%
        summarise(across(contains("num_"), sum), .groups = "drop_last") %>% 
        mutate(code_name = "Total")
    ) %>% 
    # Put the totals at the end of each group (just for neatness)
    mutate(is_total = code_name=="Total") %>% 
    arrange(course, session, is_total, code_name) %>% 
    select(-is_total)
}
produce_summary_table <- function(df) {
  df %>%
    group_by(code_name) %>%
    summarise(across(contains("num_"), list(mean = mean, sd = sd)), .groups = "drop_last")%>% 
    # Order the rows by frequency, with total at the end
    mutate(is_total = code_name=="Total") %>% 
    arrange(is_total, -num_q_mean) %>% 
    select(-is_total) %>%
    # Add the 'percentage of group' columns
    mutate(
      "num_sr_mean_pc" = num_sr_mean / num_q_mean * 100,
      "num_po_mean_pc" = num_po_mean / num_q_mean * 100,
      across(contains("pc"), \(x) round(x, digits = 0)),
      across(where(is.numeric), \(x) round(x, digits = 2))
    )
}
show_summary_table <- function(counts_by_lecture_summary_table) {
  # This is the unformatted table
  counts_by_lecture_summary_table %>% 
    kable(booktabs = T) %>% 
    kable_styling(full_width = T) %>%
    row_spec(0, angle = 90)
  
  counts_by_lecture_summary_table %>% 
    transmute(
      "Question category" = code_name,
      "Mean number of questions per lecture (s.d.)" = paste0(num_q_mean, " (", num_q_sd, ")"),
      "Mean number of student responses per lecture (as % of category)" = paste0(num_sr_mean, " (", num_sr_mean_pc,"%)"),
      "Mean number of participation opportunities per lecture (as % of category)" = paste0(num_po_mean, " (", num_po_mean_pc,"%)")
    ) %>% 
    kable(booktabs = T) %>% 
    kable_styling(full_width = T)
}

counts_by_lecture_inc_totals <- read_csv("data/output-01-counts_by_lecture_inc_totals.csv")
```


```{r}

counts_by_lecture_inc_totals_noAU <- counts_by_lecture_inc_totals %>% 
  filter(!code_name %in% c("Total", "Admin", "Understanding check")) %>% 
  add_group_totals()

counts_by_lecture_inc_totals_noAU %>% 
  write_csv("data/output-03-counts_by_lecture_inc_totals_noAU.csv")

counts_by_lecture_summary_table_noAU <- produce_summary_table(counts_by_lecture_inc_totals_noAU)

counts_by_lecture_summary_table_noAU %>% 
  show_summary_table()
```


# Variation between lecturers

```{r}
lecturer_counts <- counts_by_lecture_inc_totals %>% 
  filter(!code_name %in% c("Total", "Admin", "Understanding check")) %>% 
  add_group_totals() %>% 
  filter(code_name == "Total")

lecturer_counts_summary <- lecturer_counts %>% 
  group_by(course) %>% 
  summarise(across(contains("num"), list(total = sum, mean = mean, sd = sd))) %>% 
  left_join(anon_courses %>% select(anon_code, course_level), by = c("course" = "anon_code"))

lecturer_counts_plot_data = lecturer_counts %>% 
  pivot_longer(
    cols = contains("num"),
    names_to = "type",
    values_to ="number"
  ) %>% 
  mutate(
    type = case_when(
      type == "num_q" ~ "questions",
      type == "num_sr" ~ "questions with student response",
      type == "num_po" ~ "participation opportunities"
    )
  ) %>% 
  filter(type == "questions") %>% 
  left_join(
    lecturer_counts_summary %>% select(course, course_level, num_q_mean, num_q_sd),
    by = "course"
  ) %>% 
  ungroup() %>% 
  mutate(course = fct_reorder(course, num_q_mean, .fun='max'))
```

## Total questions asked

The number per lecture, and the mean for each lecturer, varies quite a lot:

```{r}
lecturer_counts_plot_data %>% 
  ggplot(aes(x = course)) +
  geom_point(aes(y = number), alpha = 0.5, size = 2, stroke = 0) +
  geom_point(aes(y = num_q_mean, colour = course_level), alpha = 1, size = 3) +
  scale_colour_viridis_d("Course level", end = 0.8, option = "plasma", direction = -1) +
  theme_minimal() +
  labs(
    x = "Course and lecturer",
    y = "Number of questions observed in a lecture"
  ) +
  coord_flip()

```

### Number of questions, grouped by course

```{r}

lecturer_counts_plot_data %>% 
  separate(course, into=c("course_code", "lecturer"), sep = "_", remove = FALSE) %>% 
  separate(course_code, into=c("course_year", "course_letter"), sep = 1, remove = FALSE) %>% 
  left_join(anon_courses %>% select(anon_code, anon_nickname), join_by("course" == "anon_code")) %>% 
  mutate(
    # sort by course (1A, 1B etc), then within that sort lecturers by mean number of qs asked
    ordering = 1000*as.numeric(as.factor(course_code)) - num_q_mean,
    course_and_lecturer = fct_reorder(str_glue("{anon_nickname}_{lecturer}"), ordering)
  ) %>% 
  ggplot(aes(x = course_and_lecturer)) +
  scale_x_discrete(labels = function(x) {str_replace(x, "(.*)_", "")}) +
  facet_grid(cols = vars(anon_nickname), scales = "free", space = "free", switch = "both") +
  geom_point(aes(y = num_q_mean), size = 4, colour = "#8707a6ee") +
  geom_point(aes(y = number), alpha = 0.3, size = 2, stroke = 0) +
  #geom_point(aes(y = num_q_mean, colour = course_level), alpha = 1, size = 3.5) +
  scale_alpha_identity() +
  #scale_colour_viridis_d("Course level", end = 0.8, option = "plasma", direction = -1) +
  theme_minimal(base_size = 16) +
  theme(
    strip.placement = "outside",
    axis.text.x=element_text(angle=90, vjust = 0.5),
    legend.position = "bottom",
        #panel.border = element_rect(color = "#cccccc", fill = NA, size = 1),
        #panel.spacing = unit(0, "lines"),
    strip.text = element_text(angle=90),
  ) +
  labs(
    x = "Course and lecturer",
    y = "Questions observed in a lecture"
  )
ggsave("figs/FIG_SLIDES_Questions_by_course_and_lecturer.pdf",width=30,height=15,units="cm",dpi=300)
ggsave("figs/FIG_Questions_by_course_and_lecturer.pdf",width=20,height=14,units="cm",dpi=300)

```

### Variation in spread?

Some lecturers seem to have more variable practice than others. E.g. looking at each lecturers' standard deviation of the number of questions per lecture:

```{r}
lecturer_counts_plot_data %>% 
  select(course, num_q_sd) %>% 
  distinct() %>% 
  arrange(-num_q_sd)
```

This shows that some are definitely more varied than others. These are the summary stats about the SDs shown above:

```{r}
lecturer_counts_plot_data %>% 
  select(course, num_q_sd) %>% 
  distinct() %>% 
  summarise(
    max = max(num_q_sd),
    mean = mean(num_q_sd),
    median = median(num_q_sd)
  )
```
## Types of questions

A previous table showed the breakdown of types of question asked for each course/lecturer combination. Here we highlight two examples to show the variation:

```{r message=FALSE, warning=FALSE}
all_qs_all_lecs %>% 
  filter(
    course %in% c("1E_L19", "2B_L10"),
    !is.na(category),
    !category %in% c("A", "U")
  ) %>% 
  left_join(code_details) %>% 
  tabyl(code_name, course) %>% 
  adorn_totals(c("row")) %>% 
  adorn_percentages("col") %>%
  adorn_pct_formatting(digits = 0) %>%
  adorn_ns(position = "front") %>%
  basic_kable(booktabs = T)
```

### Example generation

Looking at example-generation questions in particular:

```{r}
counts_by_lecture_inc_totals %>% 
  filter(code_name == "Example generation") %>% 
  group_by(course) %>% 
  summarise(
    X_qs = sum(num_q)
  ) %>% 
  tabyl(X_qs) %>% 
  basic_kable()
```

```{r}
all_qs_all_lecs %>% 
  filter(!code_name %in% c("Admin", "Understanding check")) %>% 
  filter(!is.na(code_name)) %>% 
  left_join(anon_courses, join_by(course == anon_code)) %>% 
  group_by(advanced, course) %>% 
  summarise(
    num_X_qs = sum(code_name == "Example generation"),
    num_all_qs = n()
  ) %>% 
  ungroup() %>% 
  group_by(advanced) %>% 
  summarise(
    num_courses = sum(num_X_qs > 0),
    num_X_qs = sum(num_X_qs),
    num_all_qs = sum(num_all_qs),
    pct_X_qs = num_X_qs / num_all_qs * 100
  ) %>% 
  basic_kable()


X_qs_details <- all_qs_all_lecs %>% 
  filter(code_name == "Example generation") %>% 
  left_join(anon_courses, join_by(course == anon_code))

X_qs_details %>% 
  group_by(advanced, course, course_topic) %>% 
  tally() %>% 
  basic_kable()

X_qs_details %>% 
  select(advanced, course, course_topic, session, time, wait_time, student_response, cluster_q, question) %>% 
  basic_kable()
```


### Questions by category

Breakdown of each lecturer's mean number of questions in each category:

```{r message=FALSE, warning=FALSE}
lecturer_category_freqs <- all_qs_all_lecs %>% 
  filter(
    !is.na(category),
    !category %in% c("A", "U")
  ) %>% 
  left_join(code_details) %>% 
  tabyl(course, collapsed_code) %>% 
  # adorn_totals(c("col"))
  pivot_longer(
    cols = !course,
    names_to = "category",
    values_to = "num_qs"
  ) %>% 
  left_join(lec_cat %>% select(course, session) %>% distinct() %>% group_by(course) %>% tally(name = "num_lectures")) %>% 
  group_by(course) %>% 
  mutate(
    mean_qs = num_qs / num_lectures,
    mean_total_qs = sum(mean_qs)
  ) %>%  
  separate(course, into=c("course_code", "lecturer"), sep = "_", remove = FALSE) %>% 
  separate(course_code, into=c("course_year", "course_letter"), sep = 1, remove = FALSE) %>% 
  mutate(
    # sort by course (1A, 1B etc), then within that sort lecturers by total qs asked
    ordering = 1000*as.numeric(as.factor(course_code)) - mean_total_qs,
    course_and_lecturer = fct_reorder(course, ordering),
    # manually reorder the categories by overall frequency
    category = fct_relevel(category, "Fact", "Next step", "Warrant", "Evaluation", "Other")
  )

lecturer_category_freqs <- counts_by_lecture_inc_totals %>% 
  filter(!code_name %in% c("Total", "Admin", "Understanding check")) %>% 
  left_join(code_details) %>% 
  group_by(course, collapsed_code) %>% 
  summarise(
    num_qs = sum(num_q),
    num_lectures = n_distinct(session),
    mean_qs = num_qs / num_lectures
  ) %>% 
  group_by(course) %>% 
  mutate (
    bar_height = sum(mean_qs),
    category = collapsed_code
  ) %>%  
  separate(course, into=c("course_code", "lecturer"), sep = "_", remove = FALSE) %>% 
  separate(course_code, into=c("course_year", "course_letter"), sep = 1, remove = FALSE) %>% 
  mutate(
    # sort by course (1A, 1B etc), then within that sort lecturers by total qs asked
    ordering = 1000*as.numeric(as.factor(course_code)) - bar_height,
    # manually reorder the categories by overall frequency
    category = fct_relevel(category, "Fact", "Next step", "Warrant", "Evaluation", "Other")
  ) %>% 
  left_join(anon_courses %>% select(course_code = anon_course, course_shortname = anon_nickname) %>% distinct()) %>% 
  mutate(course_code = course_shortname, .keep = "unused")

lecturer_category_freqs %>% 
  ggplot(aes(x = fct_reorder(course, ordering))) +
  scale_x_discrete(labels = function(x) {str_replace(x, "(.*)_", "")}) +
  facet_grid(cols = vars(course_code), scales = "free", space = "free", switch = "both") +
  geom_bar(aes(y = mean_qs, fill = category), position = "stack", stat = "identity") +
  scale_fill_viridis_d("Category", begin = 0.1, end = 0.8, option = "plasma", direction = -1) +
  theme_minimal(base_size = 16) +
  theme(
    strip.placement = "outside",
    axis.text.x=element_text(angle=90, vjust = 0.5),
    legend.position = "bottom",
        #panel.border = element_rect(color = "#cccccc", fill = NA, size = 1),
        #panel.spacing = unit(0, "lines"),
    strip.text = element_text(angle=90),
  ) +
  labs(
    x = "Course and lecturer",
    y = "Mean questions per lecture"
  )
ggsave("figs/FIG_question_categories_by_lecturer.pdf",width=20,height=14,units="cm",dpi=300)
```
<details><summary>Raw data</summary>
```{r}
lecturer_category_freqs_dat <- lecturer_category_freqs %>% 
  select(course, collapsed_code, num_qs) %>% 
  pivot_wider(
    names_from = collapsed_code,
    values_from = num_qs
  ) %>% 
  select(course, Fact, `Next step`, Warrant, Evaluation, Other)

lecturer_category_freqs_dat
```
As percentages:

```{r}
lecturer_category_freqs_dat %>% 
  adorn_percentages() %>% 
  adorn_pct_formatting()
```

</details>

The same plot but emphasising the proportions:

```{r message=FALSE, warning=FALSE}
lecturer_category_freqs %>% 
  ggplot(aes(x = fct_reorder(course, ordering))) +
  scale_x_discrete(labels = function(x) {str_replace(x, "(.*)_", "")}) +
  facet_grid(cols = vars(course_code), scales = "free", space = "free", switch = "both") +
  geom_bar(aes(y = mean_qs, fill = category), position = "fill", stat = "identity") +
  scale_fill_viridis_d("Category", begin = 0.1, end = 0.8, option = "plasma", direction = -1) +
  theme_minimal(base_size = 16) +
  theme(
    strip.placement = "outside",
    axis.text.x=element_text(angle=90, vjust = 0.5),
    legend.position = "bottom",
        #panel.border = element_rect(color = "#cccccc", fill = NA, size = 1),
        #panel.spacing = unit(0, "lines"),
    strip.text = element_text(angle=90),
  ) +
  labs(
    x = "Course and lecturer",
    y = "Proportion of questions"
  )
```

The same plot with categories in separate rows:

```{r}
lecturer_category_freqs %>% 
  ggplot(aes(x = fct_reorder(course, ordering))) +
  scale_x_discrete(labels = function(x) {str_replace(x, "(.*)_", "")}) +
  facet_grid(
    cols = vars(course_code),
    rows = vars(category),
    scales = "free",
    space = "free",
    switch = "both"
  ) +
  geom_bar(aes(y = mean_qs, fill = category), position = "stack", stat = "identity") +
  scale_fill_viridis_d("Category", begin = 0.1, end = 0.8, option = "plasma", direction = -1) +
  theme_minimal(base_size = 16) +
  theme(
    strip.placement = "outside",
    axis.text.x=element_text(angle=90, vjust = 0.5),
    legend.position = "bottom",
        #panel.border = element_rect(color = "#cccccc", fill = NA, size = 1),
        #panel.spacing = unit(0, "lines"),
    strip.text.x = element_text(angle = 90),
    strip.text.y.left = element_text(angle = 0)
  ) +
  labs(
    x = "Course and lecturer",
    y = "Mean questions per lecture"
  )
ggsave("figs/FIG_question_categories_by_lecturer_v2.pdf",width=20,height=14,units="cm",dpi=300)
```

The same plot but more greyscale-friendly:

```{r}

lecturer_category_freqs %>% 
  ggplot(aes(x = fct_reorder(course, ordering))) +
  scale_x_discrete(labels = function(x) {str_replace(x, "(.*)_", "")}) +
  facet_grid(cols = vars(course_code), scales = "free", space = "free", switch = "both") +
  geom_bar(aes(y = mean_qs, fill = category), position = "stack", stat = "identity") +
  scale_fill_grey("Category", start = 0, end = 0.9) +
  #scale_fill_brewer("Category", palette = "Purples", direction = -1) +
  #scale_fill_viridis_d("Category", option = "plasma", end = 0.9) +
  theme_minimal(base_size = 16) +
  theme(
    strip.placement = "outside",
    axis.text.x=element_text(angle=90, vjust = 0.5),
    legend.position = "bottom",
        #panel.border = element_rect(color = "#cccccc", fill = NA, size = 1),
        #panel.spacing = unit(0, "lines"),
    strip.text = element_text(angle=90),
  ) +
  labs(
    x = "Course and lecturer",
    y = "Mean questions per lecture"
  )
ggsave("figs/FIG_question_categories_by_lecturer_bw.pdf",width=20,height=14,units="cm",dpi=300)
```


<details><summary>Session info</summary>
```{r}
sessionInfo()
```
</details>
